requirements.txt
Flask==2.0.1
openai==0.27.8
Werkzeug==2.0.1
pytesseract==0.3.8
Pillow==8.3.1
python-docx==0.8.11
PyPDF2==3.0.0
python-dotenv==0.21.0
